removed and added things
